Search command is work, some extension methods

+++ src/RateMeBotVk.DataAccess/Models/Rate.cs
@@ -5,17 +5,17 @@ using System.ComponentModel.DataAnnotations.Schema;
 namespace RateMeBotVk.DataAccess.Models;
 public class Rate
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
 
     [Range(1, 5)]
     public float RateValue { get; set; }
 
-    public int RatedUserId { get; set; }
+    public long RatedUserId { get; set; }
 
     [ForeignKey(nameof(RatedUserId))]
     public User RatedUser { get; set; }
 
-    public int RatingUserId { get; set; }
+    public long RatingUserId { get; set; }
 
     [ForeignKey(nameof(RatedUserId))]
     public User RatingUser { get; set; }
